requirements.txt
amqp==2.6.1
asgiref==3.2.10
astroid==2.4.2
beautifulsoup4==4.9.1
billiard==3.6.3.0
cachetools==4.1.1
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.0
django-bootstrap4==2.2.0
django-clearcache==1.1.1
django-cors-headers==3.4.0
django-excel==0.0.10
django-filter==2.3.0
django-import-export==2.4.0
django-render-block==0.7
django-templated-email==2.3.0
djangorestframework==3.11.0
et-xmlfile==1.0.1
google-api-core==1.22.1
google-api-python-client==1.10.0
google-auth==1.20.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
isort==4.3.21
jdcal==1.4.1
kombu==4.6.11
lazy-object-proxy==1.4.3
lml==0.0.9
lxml==4.5.2
MarkupPy==1.14
mccabe==0.6.1
mercadopago==1.1.1
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.5
Pillow==7.2.0
protobuf==3.13.0
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyexcel==0.6.4
pyexcel-io==0.5.20
pyexcel-webio==0.1.4
pyexcel-xlsx==0.5.8
pyexcel-xlsxr==0.5.3
pyexcel-xlsxw==0.4.2
pylint==2.5.3
pyodbc==4.0.30
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
tablib==2.0.0
texttable==1.6.2
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.10
vine==1.3.0
wrapt==1.12.1
xlrd==1.2.0
XlsxWriter==0.9.3
xlwt==1.3.0
